Events: Expose subscribers and implement signal-slot connect.
At this point object destruction is not handled.

Declare Subscriber in metaevent.hpp.
Extend it to store a functor or a pair (object, methodId).
Store the linkId in it.
Provide a subscribers() copy accessor in MetaEvent.
Adapt connect() backend to take a subscriber as argument.
New trivial connect(object*, methodid).

Change-Id: I92d7d4e6bf0365c6a6ecf60f533a3d3234228d12
Reviewed-on: http://gerrit.aldebaran.lan:8080/2895
Reviewed-by: mnottale <mnottale@aldebaran-robotics.com>
Tested-by: mnottale <mnottale@aldebaran-robotics.com>