Fix template dllexport/import on all platforms.
Pay attention, it makes perfect sense but you need to focus:

- under gcc-macos if the template is not declared QITYPE_API, then
typeid do not match between instances created in different
libraries. This breaks Signal on Future<GenericValue> argument.
- under vcxx-win32 if the template is declared QITYPE_API, then
all instanciations that will ever be used need to be present
in the source library.
- under linux: good voodoo, it works anyway.

So the solution is to not declare the template QITYPE_API, but
to declare specific instances as QITYPE_API.

Change-Id: I077c208da01e666a2763b623dcedf1758d2f84aa
Reviewed-on: http://gerrit.aldebaran.lan:8080/6838
Reviewed-by: cgestes <gestes@aldebaran-robotics.com>
Tested-by: cgestes <gestes@aldebaran-robotics.com>