ALProxy generation.
Will likely only work from runtime data and not IDL because of expectations
on method documentation format (which the IDL spec does not handle yet).

Tries to be as close as possible from previous generator behavior, so
please do not play with indent/spacing of generated code for other
reasons.

Change-Id: I87b865c36be9e8589581e7e9429189b85fa64970
Reviewed-on: http://gerrit.aldebaran.lan:8080/17446
Reviewed-by: cgestes <gestes@aldebaran-robotics.com>
Reviewed-by: mnottale <mnottale@aldebaran-robotics.com>
Tested-by: mnottale <mnottale@aldebaran-robotics.com>