GenericObject: Bring refcounter down by using enable_shared_from_this.
The point is to hold the AnyObject while a call is being made.

Consider:

   {
   AnyObject ao(new Something);
   ao->post("method");
   }

This used to segv, since post had no way to hold a ref count, having no knowledge of
'ao'. Now it uses shared_from_this() to pass an AnyObject as call context (which
used to be Manageable*).

One bad consequence is that a GenericObject cannot be used without constructing
a shared_ptr on it. If not done, first call will fail by throwing a boost::bad_weak_ptr.

Change-Id: Ib2783bcbdb96729c994f8786095e49b6d2a94f20
Reviewed-on: http://gerrit.aldebaran.lan/23776
Tested-by: gerrit
Reviewed-by: mnottale <mnottale@aldebaran-robotics.com>
Tested-by: mnottale <mnottale@aldebaran-robotics.com>