Session: Fix a race where serviceRegistered is called too soon.
serviceRegistered is called before final registration into internal
map. So a callback immediately calling service() can end up in
a code path believing the service is remote.
This will 'work' but use a tcp connection to self, which is to suboptimal
what opening a beer bottle with a railgun is to overkill.

Change-Id: I3a6585836d07e1bd165378e6bcefcd2e68186036
Reviewed-on: http://gerrit.aldebaran.lan/29482
Reviewed-by: cgestes <gestes@aldebaran-robotics.com>
Reviewed-by: mnottale <mnottale@aldebaran-robotics.com>
Tested-by: mnottale <mnottale@aldebaran-robotics.com>