LOG: Fix bad glob checking on setCategory
Updating an existing filter with setCategory updates the value
without changing the position of the filter in the ordered list,
but it also updated existing categories as if the filter was the last element of the list.
Change that last bit by reprocessing the whole filter list
for each category that matches the updated filter.

Change-Id: I567aa78aa601644b210d5e4694805c6c1052a868
Reviewed-on: http://gerrit.aldebaran.lan/30163
Reviewed-by: hcuche <hcuche@aldebaran-robotics.com>
Tested-by: hcuche <hcuche@aldebaran-robotics.com>