add clock API based on boost::{chrono,thread}
We mimic the clocks and sleep functions from chrono and thread (in boost
and c++11) and use boost (both chrono and thread) for the
implementation.

* why do we need clocks in libqi?

We could tell people to just use boost::chrono.
But in the future, we want to enable sharing the clock epoch accross
processors and pausing the time (for simulations).
We also probably need common types for durations and time_points accross
all platforms.

* why do we need sleep in libqi?

  We could tell people to just use boost::thread, but that would not work
  for boost::thread <1.50 which has no boost::chrono support, and again,
  once pausing time is implemented, we'll need a specialized sleep_for.

* ok, but right now, why not a simple typedef for clocks?

  indeed, while pausing time is not implemented, a mere

  namespace qi {
  typedef boost::chrono::system_clock system_clock;
  typedef boost::chrono::steady_clock steady_clock;
  }

  would basically do it. However, 1) the steady_clock class also provides a
  fine place where to put the from_int32... utils 2) it enables us to use
  int64_t instead of intleast64_t as rep.

Part of redmine #17438

Change-Id: I28139b1791ce51fdfefe6585f9ddef6f0d178825