Remove TemplateTypeInterface
A template is not a type and that was very badly handled. Now template types
are registered as objects and are handled as such (useful when returning such
types over the network or to other languages). We never had problems with that
because the only cases were Future<T> and friends.

Change-Id: I2dd804197fde8064649f058814f95a2c4c0b03e3