Fix a potential deadlock in RemoteObject
When the remoteobject is closed, it unregisters itself from the socket's
messages and set all its promises to Error. However if the RO is closed
during a metacall, it can happen that the metacall pushes its promise
_after_ the cleanup, leaving it hanging. Since the RO is not receiving
any more message from the socket the promise can never be set.
This patch verifies that a close has not run during a metacall.

Change-Id: I6d7a0560ebcad7a1885a7fd435fa24976e04dc02
Reviewed-on: http://gerrit.aldebaran.lan/55927
Tested-by: gerrit
Reviewed-by: pdaouadi <pdaouadi@aldebaran-robotics.com>
Approved: tfontenay <tfontenay@aldebaran-robotics.com>
Reviewed-by: tfontenay <tfontenay@aldebaran-robotics.com>
Tested-by: tfontenay <tfontenay@aldebaran-robotics.com>