add constructors to qi::os::timeval (#1973)
This commit adds

* a constructor taking a qi::Duration, which normalizes the tv
* a constructor taking int64_t usec, which normalizes the tv
* a constructor taking sec and usec, which does *not* normalize the tv
* a default contructor to qi::os::timeval, which initializes it to 0.

Note that, following this change, this user code will not compile
anymore:

  qi::os::timeval tv = {1, 2};

it should be replaced with:

  qi::os::timeval tv(1,2);

A timeval is said to be normalized if the number of microseconds is
in [0, 1000000[.

A conversion constructor is now used in qi::os::gettimeofday, which
means the resulting timeval will be normalized.
This slightly changes the behavior of gettimeofday() for negative dates.

Change-Id: I38b3618e626bd02f75b7dba3100b871c68e978e3
Reviewed-on: http://gerrit.aldebaran.lan/56940
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran.com>
Tested-by: sbarthelemy <sbarthelemy@aldebaran.com>

add ctors to qi::os::timeval (#1973)

Change-Id: I8fd3c336230c854eaf018aabf203a0f09b4f9ddc