add qi::Clock (#1973)
add qi::Clock which comes in addition to qi::SteadyClock and qi::SystemClock
and should be used for system-wide timestamps.

The clock will eventually be monotonic, based on boost::chrono::steady_clock,
but is currently based on boost::chrono::system_clock to keep
compatiblity while all timestamps are migrated from
gettimeofday to qi::Clock.

Move the to/fromUint32 helpers from SteadyClock to Clock. (This is an
API breakage, hopefully these helpers are not used).

Change-Id: I8a1733c089c3ee3ea2015dedee45ad17a88740b0
Reviewed-on: http://gerrit.aldebaran.lan/57174
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran.com>
Tested-by: sbarthelemy <sbarthelemy@aldebaran.com>