tweak qi::SteadyClock and qi::os::ustime (#1973)
change the epoch of qi::SteadyClock to the program start time, in order
to avoid confusion with qi::Clock and to ensure it is *not* system-wise:
boost::chrono::steady_clock is system-wide, but C++ makes no such
requirement for std::chrono::steady_clock.

Also use qi::SteadyClock for qi::os::ustime, which thus becomes
monotonic.

Change-Id: I773e77895507f13a30f7a0e5c9f17f00084b782e
Reviewed-on: http://gerrit.aldebaran.lan/57072
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran.com>
Tested-by: sbarthelemy <sbarthelemy@aldebaran.com>