Removed the requirement of having to call Trackable constructor with "this".
When T inherit from Trackable<T> we are previously forced to pass
T's this pointer to the Trackable constructor in T's constructor.
Now this is optional: the default constructor of Trackable<T> will just
assume that it is static_cast-able to T and obtain T's this pointer this way.
This should help the user to avoid boilerplate.

Change-Id: I3b12882e500a75ef79a2f9109da97ec9b72167dc
Reviewed-on: http://gerrit.aldebaran.lan/58995
Reviewed-by: tfontenay <tfontenay@aldebaran.com>
Tested-by: gerrit
Approved: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>