ObjectHost: use different ids between BoundObject and RemoteObject
When passing an object through a RemoteObject pointing to a service (let's say
2), this object gets this service's id and a generated object id, on the client
side (let's say 2.3). This id may collide with an object generated on the
service side.

This causes issues for example when the service makes a call to the object and
it needs to route the response to the correct object (since two objects have the
same id on the service side).

Change-Id: I7248c3b0cd7f944429f4a384794ffaba0f03d77a
Reviewed-on: http://gerrit.aldebaran.lan/60867
Reviewed-by: pdaouadi <pdaouadi@aldebaran-robotics.com>
Tested-by: pdaouadi <pdaouadi@aldebaran-robotics.com>