Fix a bug in stream context's metaobject cache (fix for NAOqi Bug #32745).
The cache never worked because the MetaObject::operator<() was implemented wrongly.
MetaObjects are copied into the cache map and MetaObjectPrivate objects are uniquely
allocated on each copy. Comparing MetaObjectPrivate in this case does not have sense
because no 2 MetaObjects are ever equal.

Fix this by adding SHA1 hash of the MetaObjectPrivate content (function names, signal
names, property names and description are used to calculate the hash). Use this hash
to compare two MetaObjects.

Added unit tests which expose and verify that this bug is fixed.

Moved MetaObject unit test from test_object to a new file (test_metaobject) to regroup
all MetaObject tests in one place.

Change-Id: I8e3691ef2e91374f4d37491722c872c55e04ea11
Reviewed-on: http://gerrit.aldebaran.lan/66975
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: icesi <icesi@aldebaran.com>