Fix: don't continue processing method calls on unavailable objects.
This is a quickfix for a design issue.
Basically, on session destruction/disconnect, the gatway
might lose access to objects which have been called by a signal.
The original gateway code assumed that if we got a call,
then we necessarily still have a way to find the object.
This was checked by an assertion but it actually happens
and this case should abort the call attempt.

I don't think that throwing an exception is the best way to
abort the call, but I believe that a refactoring is necessary to
improve the gateway code and allow checking this case efficiently.

I tried to just return a boolean telling if or not we executed the
call, but it seems to impact all the calling code and trigger other
cases which also seem to assume that the object must be registerd,
which is incorrect.

Change-Id: I540233667cbc290507dc1772735195d7ed46afa1
Reviewed-on: http://gerrit.aldebaran.lan/68225
Reviewed-by: hcuche <hcuche@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>