Url: new, very attractive, functions
Now you can know which parts of the url are set, and set them separately.
As to demonstrate such a magnificent feature, the function specifyUrl was
written to default unspecified parts of a given url.

isValid() had to be changed a little bit so that not to accept the absurd
case where the host is empty.

Change-Id: I9f455e8174a27004af8248efb3327fe1d3191f1e
Reviewed-on: http://gerrit.aldebaran.lan/69509
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>