Add QI_ASSERT() macros.
These macros will help to customize the assertion usage in the future.
Currently, it allows to have a better behaviour with MSVC.
C's assert is only guaranteed to fail so the implementation can do
anything it wants. With GCC and clang you get nice error context so
it's ok.
With other compilers (including MSVC) there is less useful information,
for example abort can be called without any debugging information.
This is all because assert macro implementation is not defined by
the standard.

To help with that, we need to control the implementation of the
assertion macros so that we can tweak it to match our use.
This is also the beginning of allowing several useful tweaks to help
with debugging:

 - allow assertions in non-debug modes (c's assert cannot do that);
 - customize assertion behaviour (for example to log, breakpoint, then
   continue if assertions are activated in release mode,
   which can be useful to hunt
   nasty bugs that cannot happen in debug but can happen in release);

and other assertion behavior variants. It is also possible
to make it customizable on a per-process basis, which can help debugging
critical processes.

Note: This is not an alternative to throwing exceptions or error codes,
it's exclusively an internal debugging tools which help us before
releasing binaries.

This change helps with debugging the gateway.

Change-Id: I34ddafa9932fd53321efdce141c7da14a6f1cec2
Reviewed-on: http://gerrit.aldebaran.lan/68092
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmartin <jmartin@aldebaran-robotics.com>
Reviewed-by: hcuche <hcuche@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>