Equality now works even if objects travel through the network.
Until now, equality of Objects was equality of their pointers, so if an object
was sent to a service and returned, it would not be equal to the original
object anymore.
An object id is in fact a hash on the pointer passed to the object at the time
of its creation. This hash is a sha1 hash on the concatenation of the machine
uuid, the process uuid and the pointer value. An uuid is used for the process
instead of the process id to avoid different pointers at different times having
the same hash due to process id recycling.

Change-Id: I5dfa3b03bd60812b65330fb143588fdefaf71a40
Reviewed-on: http://gerrit.aldebaran.lan/71895
Reviewed-by: vpaleologue <vpaleologue@aldebaran-robotics.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>