Hotfix Gateway 3.
Gateway filters objects from messages but ignoreed when there is
a set message to a property which contain an object that could
be called later. Therefore, an Object in a property will not be
seen by the Gateway and when a call to it's members is done, the
Gateway will never find the actual object and will just send it
back to the sender (after forwarding the message all around the
world).

We now force scan messages for objects even when it's not
necessarily objects, just removed the test which predict if the
message have objects or not, it is incomplete anyway.

Change-Id: Ibf5da463b750cbd6be4d0c44460b94f849de367e
Reviewed-on: http://gerrit.aldebaran.lan/72248
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: tveltrop <tveltrop@aldebaran-robotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmartin <jmartin@aldebaran-robotics.com>