Gateway Hotfix4 : additional searches of source of objects.
When we have a complex sequence of calls of remote objects, the
gateway can lose track of which object is where. When such object is
called from multiple chain of remote objects, the gateway might not
find the destination object.
In this hotfix we look for all the possibilities before
stopping searching for objects information referenced in the gateway,
or for sockets from which the objects originate from.

Change-Id: Ib497ee3c2a01289902d1eb03037a33cb7710f989
Reviewed-on: http://gerrit.aldebaran.lan/72314
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vpaleologue <vpaleologue@aldebaran-robotics.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>