Release the mutex before executing callbacks
The future mutex was kept while executing callbacks, which may take an
arbitrary amount of time to complete.

As a consequence, a simple call to cancel() could wait for the mutex
until all the callbacks execution is complete.

Instead, release the mutex before calling the callbacks.

Related to <https://partner-portal.aldebaran.com/issues/2262>.

Change-Id: I3214b4cd01f03a43ac59e064bcdac3548aee24c5
Reviewed-on: http://gerrit.aldebaran.lan/72629
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: rvimont <rvimont@presta.aldebaran.com>