Added macros to generate Regular relational operators (==, <, and others).
Equality is central to generic value-semantic programming, so we need an
easy way to generate it. It would be nice if the compiler could generate
operator== with memberwise equality, but it is not currently possible.
Same thing for operator<.
Added also the possibility to automatically generate !=, <= and >= in terms
of == and <.
All generated operators behave correctly with each other :
== is an equivalence and < is a total ordering.
Macros are used because in the current state of the language (C++11) it is
the least bad solution.

Change-Id: I520000d1ab0ffe5577ada43bbb47194c88956d6e
Reviewed-on: http://gerrit.aldebaran.lan/73106
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>