Fixed a gateway deadlock related to object properties
As often, the deadlock was caused by the gateway redirecting the message
to the caller, a RemoteObject, that discarded it by not being able to treat it.
This fix is in fact a workaround that so far seems ok but could be dangerous :
it tries to prevent the gateway to redirect the message to its caller by
detecting that the source socket and the destination one are the same.
In this case, it searches a map of sockets populated when the harvesting fails
because the metaobject was not found, and uses the first socket found, if any.

Change-Id: Idc2cd4af8470f15e92f80ca4dfe534af513d6d1a
Reviewed-on: http://gerrit.aldebaran.lan/74151
Reviewed-by: hcuche <hcuche@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: hcuche <hcuche@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>