Fixed: dead data access through SignalSubscriber::linkId.
As a reference, the location linkId refered to was decided once and for all at construction.
This meant that, as soon as the SignalSubscriber was assigned to,
linkId was not referring to the right location anymore, but to the old one.
This resulted in modifying other objets or dead memory.
According to SignalSubscriber use in libqi, this rarely caused a problem
because we tended to copy these objects without destroying the original.

However, this is a public interface, which implies that users can
end up with random crashes by accessing linkId. So until
we change version we cannot remove linkId, but will do so soon,
as the deprecation warning already suggests.

Change-Id: I3282195a33e2c3c193c8d4b6e37d00b05f7805e6
Reviewed-on: http://gerrit.aldebaran.lan/77388
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>