TransportSocketCache: Fixed a race on socket disconnection (fixes #38167)
The context:
You ask a socket for an url from the cache. You disconnect the socket, and immediately
ask a socket for the same url from the cache. Depending on the race, you can
end up with a disconnected socket, instead of a connected one.
The race is between the disconnection of the socket and the receiving of the
`disconnected` signal by the cache. If this signal arrives too late, the socket
is still in the cache when we ask it, and it is returned although it is
disconnected.

Change-Id: Ife655f7b2c8d8b6c5eddc08b801d1f41ac8f1b52