Type: fix crashy conversion of AnyReference from invalid values #37978
The tests revealed many intermediary issues that are only fully fixed
with the last patch in AnyReference:
- the setter defined in the Object type interface had an undefined behavior
when receiving dynamic values with invalid content. Now, the setter throws
if an invalid value is encountered in the source.
- conversion function from AnyReference to object did not handle exceptions
so behaved differently from other conversion functions.
- AnyFunction::call did not properly check the arguments before using them,
and in fact did not support invalid values cleanly. Now, conversion
failures are properly reported. Documentation is updated accordingly.
- Pretty signature of invalid values is set to "Invalid".

Change-Id: Ic630fb714865a2895f495d8f6847dcc61213e19b