Generic Object: meta call to track the underlying future properly #38551
GenericObject::metaCall's returned future did only track the underlying
call being performed, and did not adapt the future to the underlying call's
returned future. It is now fixed and specified explicitly in the docstring.

This change required that adapting futures of AnyReference was made
possible by changes in futureAdapter, which naturally involve an additional
copy.

The commit include tests proving that it does not work in the direct mode.

Change-Id: I29c3b02a7bd29a5617129794ed6fb229111a1fea
Reviewed-on: http://gerrit.aldebaran.lan/82900
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>