socket: Makes connecting state cancelable by short-circuiting the connecting steps. (fix #38450)
In the previous version, the code was taking into account a disconnection request only on state change.
For instance, it waited the end of the connection to effectively start the disconnection steps.
The goal was to have the simplest and most predictible code flow (no shortcut, always the same
succession of states).

It happens that if the server crashes during the client connection, the connection will have to wait
a system timeout to end, and then the disconnection request will be processed. This timeout is
often too long to be acceptable.

In the new version, requesting a disconnection during the connection steps will provoque an immediate
abort of the connection. This solution is clean, because the code flow, in terms of state, does not
change. A cancelled connection will result in the same further states (disconnecting, disconnected).

Change-Id: I34dfdb7438b35b313faea68ef2d46fd136c6d254
Reviewed-on: http://gerrit.aldebaran.lan/83000
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>