Fixes QI_NOTHROW macro implementation.
The previous implementation was relying on the C++ macro telling
the supported C++ version. However this is not enough to know if
a specific feature is supported or not by a compiler and should
never be used to detect features. There are other macros in C++14
to check specific features, but the simplest way to do it is to
just rely on Boost's implementation of the same macro.

Change-Id: If16f2d98f46739ecfc84ca6ba87bc69c84935d81
Reviewed-on: http://gerrit.aldebaran.lan/83349
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>