Fixes qi::bind compilation: Workaround VS2015 compiler bug.
VS2015 compiler (vc14.0)  might have an incorrect behaviour in expressions mixing `decltype`
feature with either template typedefs (with `using` keyword) or tail return type
which are (all 3) features they implemented recently (but are fixed mostly in vc14.1/vc15, VS2017).

The test added here reproduces one case found by client code which put in evidence this
compiler bug.
The fix is to try to express the same idea but with different features,
here using a template type in a non-template typedef using (instead of a template typedef).

Change-Id: I8527c52b7d43b39decb2f19f5f54cf50b958f89e
Reviewed-on: http://gerrit.aldebaran.lan/83465
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>