tests: Corrects several tests in gateway mode
Some tests would fail when launched in gateway mode
because they would try to access a service right
after registering it, which can fail because it takes
some time for the gateway to mirror it.
It also takes some time for the gateway to relisten
after being disconnected then reconnected to the service
directory.
To solve this, this patch adds a few waitForService and
sleepFor.

Change-Id: I71f7cfe51abf3da9efcf9a7ddfeba882f12a9bbe
Reviewed-on: http://gerrit.aldebaran.lan/84432
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>