messaging: lock socket state mutex before disconnecting
Creating a DisconnectingState closes the socket as a side effect.
A race condition would occur if between the time the state was
created and the _stateMutex of the TcpMessageSocket was locked
TcpMessageSocket::send was called. That call would lock the mutex
and check the socket state which would not have been changed yet
although the socket would still be closing. This would lead to a
crash in asio.

Change-Id: I4938cb2bc4529a83f3c2a3ac5d73f65f1618525e
Reviewed-on: http://gerrit.aldebaran.lan/84750
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>