event loop: Fixes total tasks counting and stops io_service
Some std::atomic were not initialized and had undefined values,
they are now initialized to zero.
The behavior of counting the total tasks was poorly implemented,
it has been reworked to be more robust.
The io_service was in some cases not stopped when _work was
destroyed, it is now stopped explicitly.
As EventLoop::start can be called after stop, _work is now
instanciated in the start method, and the io_service is also
reset.

Change-Id: I33f32402a39d96ef105e7c1ae4f0237667a46062
Reviewed-on: http://gerrit.aldebaran.lan/85084
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>