tests: remove tests from sessionpair that did not use it
Some tests in test_send_object, which is launched as part
of test_messaging_with_sessionpair did not actually use
TestSessionPair and as such did not respect the mode that
was given to it.

These tests are moved into separate files that are launched
as part of test_messaging instead.

Change-Id: Ia1206a1df5327ccfb9a008c5d78af47ddcb81fd7
Reviewed-on: http://gerrit.aldebaran.lan/85724
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>