messaging: Fixes data races in SetupConnectionStop
In the call operator that takes a socket, the call to close()
needs to be wrapped in the io_service strand.
In the call operator that takes a resolver, the lifetime of
the resolver could not be enforced by the caller at the time
the continuation to stop was called.
SetupConnectionStop now takes a syncTransfo and a lifetimeTransfo.

Change-Id: I942e312978162b4dc1b7a3c4534291c34f91f2fe
Reviewed-on: http://gerrit.aldebaran.lan/85740
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>