messaging: Transforms connecting/connected result to shared_ptr
There were cases where the result of the connecting/connected
state could be set to the complete promise before being
modified (eg by a call to stop on the socket), since there can be some
time between the moment we set the complete promise and
the moment the continuations are called which is when
the state is effectivily changed.

Results are now treated as entities and are kept inside
a shared_ptr, protected by a synchronized_value.

Change-Id: Ibece9eaf212e4ca320f00f367cec7aed704ce9c9
Reviewed-on: http://gerrit.aldebaran.lan/85738
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>