Signal Spy: allow waitUntil to be cancelled
The fix required the addition of a new structure. Since the structure is
not relevant for clients, it was put in a new .cpp file. By the way,
everything that could be moved to the .cpp file was moved, improving
compilation times and binary sizes.

Also, all useless qi:: prefixes were removed, and header guards were fixed.

Change-Id: I3b5a72749e87c257d8a7834d0d2f819d56606319
Reviewed-on: http://gerrit.aldebaran.lan/86274
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vpaleologue <vpaleologue@aldebaran-robotics.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>