Fixes Session::waitForService blocking call.
Session::waitForService call must never block (the user can wait on the
returned future, or not) but it was calling Future::hasError
which is a blocking call.
If the caller get the future and do not wait but do not intend
to wait for the result, the call would block anyway, making
the future useless.

Change-Id: I7156fb2075ddbe73b3e3b0cb9861787b1e9b04ae
Reviewed-on: http://gerrit.aldebaran.lan/86518
Reviewed-by: gvallat <gvallat@aldebaran-robotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>