messaging: Ensures that the future returned by `Session_Service::service()` is always set (#40203).
Previously, the code setting the promise could be skipped because of a
premature return or an exception.
RAII is now used to ensure this is not the case.

Still, the future returned by `Session_Service::service()` could be never set
if a distant component never answers.

Also, verbose logs have been added to facilitate debugging.

Change-Id: Ie6daab665b92eace224fe99560bc90f19d9fd253
Reviewed-on: http://gerrit.aldebaran.lan/86596
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>