messaging: Adds a timeout to `Session::waitForService()`.
This is necessary to avoid potential infinite waiting.

The historical version (taking just the service name) is redefined to use a
default timeout. There is also now an overload that explicitely takes the
timeout.

Due to various system considerations, the default value for the timeout is for
now set to 5 minutes.

Change-Id: I73e8e974e17d69b1dd78f2030a9b2606e5462f2a
Reviewed-on: http://gerrit.aldebaran.lan/86610
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>