tests: Removes dependency to libqi for execs used in "spawn" functions tests.
Spawn functions tests (including ScopedProcess usage)
on Windows can fail because the launched executable fails on
startup as the working environment is incorrect (the executable is in
a temporary directory with no other dependencies)
and the system does not find boost libraries dlls
which are required by libqi.
This is because we do not make the child process inherit the test process environment
when calling the spawn functions.

These executable do not need libqi anyway so removing the dependency
fixes this issue (they can be run standalone).

Change-Id: Id88b12c0a9fe7ad0328e9ee65e0707f76b13918e
Signed-off-by: Joël LAMOTTE <jlamotte@softbankrobotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/87633
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>