tests: Fixes tests not passing because connecting too fast to child process.
Sometimes (mainly on Windows in Debug mode) a child process will take some
time to startup. When a test launches such process and then tries
to connect via a socket to it, a connection error might happen which might make
the test fail while the issue is not related to the test.
To prevent this, we provide test::tryConnect() to attempt several connections
before considering the connection to be failed.

Change-Id: I2a9bdda447d18e80d0d080c1525d8fcab2cdaf40
Reviewed-on: http://gerrit.aldebaran.lan/87996
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>