messaging: Decorellates Disconnecting state construction from socket closing
Disconnecting state construction also used to schedule a close
on the socket, which meant that we had to lock the _stateMutex
of the TcpMessageSocket during that construction and until
the _state member of the TcpMessageSocket was set which was unwanted.

This patch adds a call operator to Disconnecting which triggers the
close. This allows us to create the new state outside of the lock (since
it has no side effect anymore), lock the mutex, set the _state variable,
unlock the mutex and then only trigger the close.

Change-Id: I3570c44b0d59f68b6afbcb6ad34b1496f7c5817d
Reviewed-on: http://gerrit.aldebaran.lan/88217
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>