Object: Makes conversion to boolean explicit.
This is to avoid unwanted conversions from `Object` to `bool` and thus get a
better type-safety.

Implicit conversion to `bool` was unintendly used in unit tests, specifically
in the expressions `ostream << obj`. Despite the lack of `operator<<` for
`Object`, this compiled because of the implicit conversion from `Object` to
`bool`. A dummy `operator<<` for `Object` has been added in unit tests to
compensate (see `objectio.hpp`).

Change-Id: I041727375899e597709d94954b1a53980d61ef42
Reviewed-on: http://gerrit.aldebaran.lan/89234
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>