ptruid: Changes definition of `end()` to avoid dereferencing an invalid iterator.
Context:

`PtrUid` stores data as a `std::array` of `uint32_t`. This makes copying,
comparing, and other operations more efficient than if a `std::array` of
`uint8_t` was used.

Nevertheless, it is sometimes useful for the user to be able to iterate on
bytes rather than on `uint32_t`s. `PtrUid` therefore also provides `uint8_t`
iterators. The definition of the `uint8_t` version of `end()` reused the
`uint32_t` version and was implemented as:
`static_cast<uint8_t*>(&*endUInt32())`.

It is not legal to dereference an 'end' iterator. In this case it is unlikely
to cause a problem since it should be optimized by the compiler. But it does
make debug tools complain.

The implementation of `uint8_t` `end()` was changed to: `begin() + offset`.

Change-Id: I437a4b263bbdeffda4af80862201021b2b82c470
Reviewed-on: http://gerrit.aldebaran.lan/89457
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: matthieu.paindavoine <matthieu.paindavoine@softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>