Strand: clarify schedulerFor and clean the handling of arguments
Several tests were added according to my current use. It seems using
std::bind with std::ref on a stranded function does not react as I
expected, but these tests made clear that this is the normal behavior.

Documentation was added to help the next ones that may play with this
function, as well as for the users of unwrappedSchedulerFor.

Change-Id: I846077e6ff1d2f8567d0cddbf55b2bb53f2e9d6d
Reviewed-on: http://gerrit.aldebaran.lan/89956
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>