future: Adds safer alternatives to `value()`.
`value()` returns a reference to a constant.
As such, this reference must not outlive the `Future` it comes from.
It is however easy in compositional code to end up with a dangling reference.

New safer alternatives are now:

- manipulating a copy of the value, returned by `valueCopy()`

- extending a `Future`'s lifetime by manipulating a `shared_ptr` returned by
  `valueSharedPtr()`

The original `value()` is neither changed nor deprecated, as it can be more
performant, provided the user takes the responsibility of the lifetime issue.

Change-Id: Iaea9b1a2517a02a3dc0bb70e67f70ee36836f2ba
Reviewed-on: http://gerrit.aldebaran.lan/90128
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>