messaging: Adds capability "ObjectPtrUID" for protocol retro-compatibility.
Without this capability, a process that was not compiled with PtrUID
but communicating with a process compiled with PtrUID will receive
messages with unhandled data that it will still attempt to read and
will end with incorrect data.

Adding the capability makes both process know if the ptruid will be
sent and receive or not. When a process A can handle ptruid but not another
process B, A will create a new ptruid for the received object.

Change-Id: Iefed74ac95b2cf3d03b3ac7d61f37f0fe81212d0
Reviewed-on: http://gerrit.aldebaran.lan/90191
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>