ka.scoped: Removes move-assignment.
A correct move-assignment must call the function of the target `scoped_t`
before performing the assignment by itself. This is not currently done.

Doing so means that a default-constructed `scoped_t` that is moved to will have
its function called. But there is no way to know if this default constructed
function is valid or not (the `Regular` concept does not give this guarantee,
as it relies on the notion of 'partially-formed state' for default-constructed
values).

Solutions do exist to distinguish valid default constructed values from invalid
ones (by introducing a tag type, etc.), but they hurt the design simplicity of
`scoped_t`.

Consequently, move-assignment is now deleted. Users that want to
default-construct a `scoped_t` must wrap it in an optional.

Change-Id: I471d0db9d191975572ed6c9f5b94f64e53ec90b2
Reviewed-on: http://gerrit.aldebaran.lan/90640
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>