cmake.tests: Installs libtestsession even with QI_WITH_TESTS=OFF.
Some projects use libtestsession for their own tests while using libqi
from the toolchain. The package qi-framework which includes libqi
is compiled without the tests (QI_WITH_TESTS=OFF) when it's compiled
for the toolchain which prevents having libtestsession available in
the toolchain package.

This patch is a workaround to that problem. A more long-term solution
would be to move libtestsession and all it includes inside libqi.

Change-Id: Ib104a7909cfb419f3b35f74bc016237a60873a76
Reviewed-on: http://gerrit.aldebaran.lan/90672
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>