genericObject: Plugs memory leak.
Ensures that the shared_ptr is reset upon destruction of the
corresponding GenericObject. Replaces original function by a lambda
because it was only called in one place.

The memory leak occured because the previous deleter kept a copy of the
shared pointer of the Proxy. If the deleter was kept alive by a weak
pointer (eg WeakObject), then the shared pointer may never be
destroyed. The lifetime of the shared pointer of the Proxy is then
linked to the lifetime of the deleter and thus linked to the last weak
pointer lifetime, which is incorrect.

Fixes #41451

Change-Id: I4f1b5f9cf3de64af42837ca884fa93c1927e9e57
Reviewed-on: http://gerrit.aldebaran.lan/91300
Reviewed-by: matthieu.paindavoine <matthieu.paindavoine@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: matthieu.paindavoine <matthieu.paindavoine@softbankrobotics.com>